Revert claiming support for NC21
Although the app was already working fine with Nextcloud 21 beta5, the
compatibility was broken again in beta7. This was caused by DBAL
exception no longer flowing through to the applications, as explained
in https://github.com/nextcloud/server/issues/25415.

The issue linked above has to be solved one way or another in Nextcloud
server before the Music app's support for NC21 can be restored. Still,
it will require changes for the app, too, after the server-side fix.