Fix playlist and radio exporting not working on NC versions after 13
There has always been a small API difference on the filepicker between
ownCloud and Nextcloud, but this was not taken into account in the
Music app. However, the use of the picker had worked correctly simply
by chance (*) up until Music 1.0.0 added some argument type declarations
on the back-end. Coupled by the activation of the strict typing in NC14,
this made the bug very visible and broke the playlist exporting
functionality.

(*) We were passing integer 1 as argument `oncollision` while it was
supposed to be the string "abort". However, any unknown argument
values had the same handling as the string "abort", hiding the bug as
long as no-one was checking the argument types.

refs #822