Subsonic: Fall back to xml if jsonp format requested without a callback
The return Subsonic responses in jsonp format, the client has to pass
another argument `callback`. Previously, we blocked the call if this
argument was missing, but now we just fall back to using the xml
format. This is what the real Subsonic server does in this case, too.

When the Jamstash client is used in the (default) jsonp mode, it sends
its requests for album art and audio streams without giving the
`callback` argument. This might be a client bug but as this works on
the real Subsonic server, it's better to have it working on the Music
app, too.

refs #787