Subsonic: Support also updating existing lists with `createPlaylist`
According to the Subsonic API specification, the method
`createPlaylist` can be used also to modify the contents of an existing
playlist in addition to creating a new playlist. When modifying, the
client has to pass the argument `playlistId` instead of `name`.

Renaming the list with the method is not supported, even if both
`playlistId` and `name` are supplied. This is not clear from the
documentation but was cross-checked from the implementation of the real
Subsonic server.

The Jamstash client uses this mode of the `createPlaylist` method when
rearranging existing playlists.