Update the translation template to match the latest software
The html for the "Getting started" of the Internet radio had to be
reformatted to extract a bit nicer translation keys; the tabulators used to
format the file got originally extracted as part of the translation keys.

One obsolete translation was removed from fake-template.php. Nowadays,
the embedded Files player actually contains a lot of translatable strings
which are not present in the fake-template. We should really create some
better system to extract those, so we don't try to add those to the
fake-template, at least for now.