Tweak the tooltip of the play icon on the navigation pane
Because of the way the string has been defined in the template, the
localization extractor actually can't find this localizable string. But it
doesn't matter, if we use so generic string that it has been extracted from
elsewhere. It can also be argued that the previous longer string "Play list"
wasn't adding any value compared to the shorter version "Play".