Support for Nextcloud 21
NC21 wraps all the DBAL exceptions to its own generic exception type. This
makes it a bit trickier to react to unique constraint violations. To not have
to write the cumbersome try-catch-if-else blocks everywhere, both the native
DBAL UniqueConstraintViolationExceptions and the wrapped NC21 exceptions are
now caught as early as possible, and mapped to our own exception type.

refs #830