Fix a bug introduced by ab088210 on the disk number migration
Changing the exception type had been accidentally applied also on one
occasion where it shouldn't have been done. Because of this, updating
from Music version <0.13.0 to 1.0.3 may lead to corrupted database on
some cases. To recover, the music collection has to be wiped clean and
rescanned.

Hopefully, not many users are still using the over a year old versions
<0.13.0. And if they are, then hopefully they don't update  now to
1.0.3 (without first installing some version in between).