Travis: Hack to enable testing on top of Nextcloud master branch
Normally, the Nextcloud master branch has the version number of the
next major release of NC. We don't want to declare compatibility for
that major version in the Music app, because incompatible changes may
appear to that branch in which ever moment. The contents of the next
major release start to stabilize only upon the PR releases, but at that
point, the development moves to the stable branch, and master starts to
track again the next major release.

Yet, we want to run automatic tests on the master branch to get some
heads up when the compatibility breaks. To achieve this, we now
mutilate the info.xml file on Travis before activating the Music app,
to claim support for (almost) any Nextcloud version.