Support sorting playlists by title, album, or artist
The actions menu of the playlists in the navigation pane now contain a new
action "Sort …". This opens a sub-menu with three options: "by title", "by
artist", and "by album". Selecting one of these sorts the playlist in
question according to the selected property, and syncs the change to the
server.

refs #689