Fix playlist "updated" timestamp not updating on the UI on some modifications
The "updated" timestamp of a playlist was not changed on the web UI when a
track was removed from the playlist or the tracks were manually reordered
with drag-and-drop. The underlying timestamp on the server end, on the other
hand, was updated and the change was propagated to the UI upon the web page
reload.