Fix unnecessary view scrolling when opening track details in the playlist view
The generic logic in maincontroller.js can't detect if the track in question
is actually in the view port in case of the playlist view. This is because
the view contains no elements having element IDs like "track-<id>". This is
because the track IDs are not necessarily unique in the playlist view.
Previously, the scrolling got triggered every time when the searched element
didn't exist, but now we bail out with a null check.

It's theoretically possible, that the track under investigation would scroll
out from the view port now when viewing details within the playlist view. This
should be rare enough, though.