Add tooltip for the truncated titles in the details pane
The tooltip shows the full version of the detail title when it can't
fit the provided space.

Adding the tooltip only when the text is truncated was not quite
simple. In the end, a solution inspired by
https://stackoverflow.com/a/60073230/4348850 was selected. The other,
neater looking solutions on that SO article did not work in our case.