Fix playlist comment modification not always synced to server
The changes made on the comment were not pushed to the server in case
the user clicked the comment text box again after the modification but
before committing the changes. This was because the logic to start
editing got triggered again, and the `initialComment` got set to the now
changed value. Then upon committing, the logic evaluated that the
comment has not changed and doesn't need to be pushed to the server.