Show only one desktop notification at time
The previous desktop notification is now always closed when a new one
is opened. On desktop browsers, this makes a difference only when the
playing song changes in rapid succession, and the popup has no time to
close automatically before the new one is presented. On Firefox for
Android, all the notifications stay in the notification center
indefinitely, and this change prevents them from piling up.

Also, debounce of 500ms is now applied before showing the notification.
This is because programmatically closing the notification doesn't work
right after it has been spawned. Hence, rapidly tapping the "Skip next"
button still made the notifications pile up.

refs #828