Show play icon on hover next to the album title
This replaces the play icon overlayed on the album art when hovering
the album title. The overlayed play icon can not be used in the
compact albums view, and it had the technical problem that it was
visible also when hovering the album info button. Also, the overlayed
expand/collapse icon in the compact view was visible while hovering the
album title, and this was plain wrong and misleading. The overlayed
play icon on album art is now shown only when hovering the album art
itself.