Modify play icon overlay to be clearly visible on all backgrounds
The play icon overly shown when hovering an album cover in the Albums
view now has a "back glow". This makes the icon stand out even when the
album cover on the background is (almost) the same color as the play
icon (i.e. very dark on the default theme or very light on the dark
theme).