Fix details icon not being shown after a truncated album title
In the Albums view, the details button was not shown on those albums
with name so long that it got truncated. The problem was worse on the
"compact" layout as more album names get truncated there. The solution
should work at least on the sufficiently modern browsers.

We no longer allow the album titles to be wrapped to multiple lines
even on the mobile layout, as it was interacting badly with this fix.