Playlist files: Support for backslashes as directory seprators
When parsing a playlist file, we now support also the Windows style of
using backslashes as directory separators instead of slashes. The
solution is kind of heuristic or "best effort": The parsed path is
first tried to be used "as is". If there is no such file and the path
contains any backslashes, then those backslashes are converted to
slashes and the lookup is made again. It is principally possible, that
this would sometime find some unintended file, but those cases are
"pathological" and should not be a real-life issue.

refs #845