Make the Albums compact layout scale better to the window width
Fixed three issues with the compact layout of the Albums view on
"tablet" and "mobile" screens:

1. On relatively wide "tablet" screens, only one column of albums was
being shown even though two columns could easily fit the screen.

2. On a bit narrower "tablet" screens, when there was no longer room
for two columns, the album and track titles were still truncated to
much shorter length than necessary, leaving a lot of unused space.

3. On "portrait mobile" screens, the album names were not truncated to
the available space, overlapping with the alphabet navigation control.

The key to all of these issues was to separate the evaluation of
"single-column" mode from the evaluation of the "tablet"/"mobile" mode.
The screen width limit to switch to the "single-column" mode is
different when using the "compact" layout instead of the "normal"
layout.