Ampache: Return all results if argument `limit=0` given
This is how the "real" Ampache server works, too, although this is not
documented in the current API specification for v4 or v5 API. The Amarok
player uses this argument value, and there should be no real-life case where
the client actually would want to have 0 interpreted as literal value of the
limit.

Furthermore, the Music app already tried to interpret zero as "unlimited" but
this worked only with integer 0 and not with string "0". Whether the numeric
arguments come as strings or integers, depends on the details of the hosting
cloud, and may differ between versions of ownCloud and Nextcloud.

refs https://github.com/owncloud/music/issues/854