Enforce the internal encoding in PHP to be UTF-8
It is possible that the internal encoding used by PHP is something else than
UTF-8, depending on settings given in php.ini. Still, the OC/NC database
should always be in UTF-8 and the getID3 library returns UTF-8-encoded data
by default. Hence, using internal encoding other than UTF-8 may lead to
unexpected results in many places, especially when any of the PHP built-in
functions `mb_*` are used. To avoid this, we now override the internal
encoding to be UTF-8 in the constructor of the main application class.

refs https://github.com/owncloud/music/issues/846