Don't require `allow_url_fopen` to scan music files
Eight years ago, there may have been some reason to disable scanning if the
php.ini file didn't have the option `allow_url_fopen` enabled. However,
nowadays there doesn't seem to be any reason for that; the scanning works
just fine without it. Also, it's pointless to try to enable it within the php
code because that will never work because of the security reasons.

Where the `allow_url_fopen` is actually needed, is to fetch the data from
Last.fm in case the integration is enabled. But even there, there's no reason
to check the status of the setting separately because fetching the data will
just fail and emit an understandable error message to the log which will guide
the admin to the root cause.

refs https://github.com/owncloud/music/issues/763