Clean up some style issues reported by ESLint
Also, configured ESLint to use newer ECMAScript version (2021) as we
nowadays use a small bit of ES syntax newer than ES5 (Babel plugin then
converts this to ES5 compatible code during the webpack build).

Finally, set the sourceType of ESLint to "module" so that it doesn't
complain about the use of the 'import' keyword.