Stream the audio files so that they don't need to be allocated to RAM
Previously, we were returning audio files by first reading them to RAM
as whole before sending them to the client. This could cause problems
with large enough files (usually FLAC or WAV), as the `memory_limit` in
php.ini could prevent such memory allocation.

Now we, instead, copy the file data to the output stream lazily. This
can be achieved by implementing the interface ICallbackResponse. The
cloud core would also provide a ready-made implementation
StreamResponse, but that wouldn't support the range requests.

It's possible that the stream-like responses also improve the
performance with external storages. However, this hasn't been tested
and it may also depend on the implementation details of the storage
plugin used.

refs #864