Show the play icon overlay on album covers also in compact search mode
When using the albums view compact layout, the function of clicking the album
cover is different normally and while in the search mode: Normally, it
expands the album, but during the search, there's nothing to expand and
clicking the cover plays the album. This is now indicated also by showing the
play icon overlay on hovering the album cover, just like in the "normal
layout" mode.

To make things simpler, all the margin needed for the play icon overlay is
now part of the icon, just like with the "expand" overlay icon. To facilitate
easier modification of the icon using a text editor, the useless translation
to some random coordinates was refactored out.