Fix performance problem on Subsonic actions `getAlbumList(2)`
When getting a large list of albums with the Subsonic actions
`getAlbumList` or `getAlbumList2` (up to 500 albums supported), the
SubsonicController used to fetch the Genre objects separately from the
database for each album. These objects were used to obtain the genre
names contained on the album, but contained also other information like
the track/album/artist counts of the genre, and these were not trivial
to calculate. These DB queries could spend several minutes with 500
albums, unless the DB query caching could save the day.

This has now been optimized by gluing also the genre names to the Album
objects already when fetching those albums from AlbumBusinessLayer. We
utilize here the Genre class without setting all its properties, so
this is something to be aware of when using those Genre objects. There
is the risk of tripping on this in the future, but I didn't want to
create another "PartialGenre" class just for this one case.

The same optimization should improve also the performance of all the
other Subsonic actions returning albums (e.g. `getAlbum`, `search2`,
`search3`), but the relative effect is not so dramatic when the number
of albums returned by a single call is smaller.

refs #873