Fix Last.fm tag/link not showing correctly if there is only one
In case the metadata retrieved from the Last.fm service contains only
one "tag" or "similar to" field, this field is not presented in the
JSON result as an array but as an individual object. When this was not
taken into account, the UI showed in such cases two links with the text
"undefined" and URL "undefined".