Move Subsonic API converters to PodcastChannel and PodcastEpisode
The entity classes PodcastChannel and PodcastEpisode now have functions
called toSubsonicApi. These replace the former functions
SubsonicController::podcastChannelToApi and
SubsonicController::podcastChannelToApi. Similar design was previously
applied to the Ampache API conversion of the podcast entity types. The
target of this refactoring is to reduce the size of the monolithic
SubsonicController.

Similar refactoring should be eventually made to the other entity
types, too, but that does not belong under this feature branch.