Subsonic: Include properties `suffix` and `bitRate` on podcast episodes
These pieces of information are not directly available in the RSS feed
but they can usually be evaluated from the available information. The
major motivator for this improvement was that the Jamstash client
depends on the `suffix` field to be able to play the episode at all.
Also, Ultrasonic client showed an ugly looking "-1 kbps" on each
episode while there was no bitRate available.