Add option `--force` for the command `occ music:podcast-update`
The option may be give to force the updating of the episodes of all the
targeted podcast channels, even when the podcast feed has not changed at all
since the previous update. This may be desirable during the development when
the parsing of the feed is changed by any means. It might be useful for the
end-user, too, in case the database update somehow got unexpectedly aborted
during the previous attempt.