Add podcast episodes to DB in chronological order
When subscribing to a podcast channel, add the available episodes to the DB
in reverse order compared to the order of the RSS feed. Typically, the most
recent episode is the first in the list, so this means adding the episodes
from oldest to newest. Then, when new episodes are published, they will
obviously be added after the first patch of the episodes, maintaining the
same "oldest to newest" logic. Then, when returning the episodes through the
REST API, the list is reversed once more to get the newest episodes first.