Subsonic: Support method `getNewestPodcasts`
It's not quite clear if this method should return the episodes according to
their original publish dates or by the order in which they have been added to
the database. Now it's doing the latter which was a bit easier to achieve.

The Subsonic API version has been bumbed to 1.13.0 to advertise the
availability of this method; at least DSub uses only methods belonging to the
claimed API version. The other new methods from v1.12.0 and 1.13.0 are
`getTopSongs`, `getPlayQueue`, and `savePlayQueue`, which have been stubbed
for now.