Fix the layout of the podcast channel actions menu on ownCloud
NC core provides suitable styling for this but OC core does not. Hence,
we need to define the style on our own. The correct kind of styling
already existed in style-navigation.css but that was applied only
within the actions menu. These style definitions were now moved to
their own file style-actions-menu.css, and they were made to apply both
on the navigation pane and elsewhere within the Music app.