Fix flickering on the details pane
When the details pane (for any content type) was closed and then opened
again for any content type requiring closing, there was a flicker when
the previous content was visible on the screen for a fraction of a
second before being replaced by the loading spinner. This was visible
at least when opening details pane for a podcast channel/episode, a
track, or an album.