When updating/deleting a channel, show a load spinner next to its title
Also, when creating the first channel, there is now a loading spinner
shown in the view.

To facilitate all this, all methods of PodcastService were modified to
return promises. This replaced signaling via $rootScope.$emit which
was really designed only to be used when triggering operations from the
navigation pane actions menu.