Ampache: Use ISO 8601 format for all podcast timestamps
The same format is used also on the response of the Ampache action
`handshake`. The example responses in the Ampache API documentation
actually use the American format like "08/19/21 22:53 PM" but I hope
that this is just an error in the spec and/or in Ampache SW since using
ambiguous format like that in any API makes no sense.