Extract and store lastBuildDate from podcasts after all
This timestamp is part of the specified Ampache podcast responses so we
need it after all. It is now also included in the channel details shown
on the web UI. This field is still excluded when judging if the RSS
feed has actually changed after the previous podcast update (because on
some channels, the lastBuildDate changes very ofter without any other
changes). This means that if only the lastBuildDate is updated in the
feed, then the user will not see this even with manual update from the
UI.