Initialize the "excluded folder" picker to the library base path if possible
When opening the folder picker UI to set an excluded folder, the dialog is
now opened with the current music library root folder as base path. However,
the filepicker dialog component supports this only on Nextcloud and only
starting from the version NC16. That is, on ownCloud and older versions of
Nextcloud the folder picker UI always start from the user home folder as
before.

refs #876