Subsonic refactoring: Use reflection to inject request parameters
All the methods corresponding the Subsonic REST API methods now declare the
required and optional request parameters as parameters for the function.
Required parameters are type declared with string, int, or bool while
optional parameters are type declared with ?stirng, ?int, or ?bool. Repeatd
parameters are type declared as array.

The handleRequest method then uses reflection to extract the corresponding
request parameter values and inject them to the handler method call. When a
parameter has not been declared as nullable, it is interpreted to be a
required parameter and an exception is thrown if there no such parameter in
the HTTP request. Non-null parameters are casted to the declared type.