Ampache refactoring: Dispatch API calls based on method signatures
AmpacheController now dispatches all the API calls using similar
mechanism as SubsonicController: Reflection is used to find the method
implementing the API action in question, and also required and optional
parameters are extracted and injected according to the method signature.

With this, we get rid of the huge switch-case within the function
AmpacheController::dispatch, where it was difficult to see if all the
positional parameters were passed correctly. Also, there is now only one
system to pass request parameters to the action handlers while
previously "commonly used" parameters were passed as method arguments
and parameters specific for one or a few actions were obtained with
$this->request->getParam(...).

This commit includes also a small functional change: All Ampache result
are now limited to 5000 entries at maximum, even if no limit or larger
limit was passed by the client. This is required by the Ampache API
specification.