Ampache: Optimize rendering of larger sets of songs
All the albums corresponding the tracks are now fetched with a single
DB query. The effect is significant when rendering a thousand or more
songs. Logic for this already existed in SubsonicController, but it was
now moved to AlbumBusinessLayer to reuse it with Ampache.

With this change, the special handling for the action `songs` without
any limiting factors (i.e. fetching all the songs) became unnecessary.
Actually, this is significantly more efficient than the previous
special handling, mostly because `usort` is no longer user in this case.

Also the small optimization of `album_songs` now became irrelevant.