Ampache: Optimize `playlist_songs` when called for the 'All tracks' list
Sorting the result set within the DBMS is significantly more efficient
than usng usort in PHP. On the 2000 track test library this reduced the
response time of the action from ~4.6s to ~2.6s. The effect is
significantly larger if pagination is used: The previous implementation
fetched and sorted *all* the tracks of the user even if only a small
slice was requested.