Use case-insensitive matching in findAllByAlbum even when $fuzzy=false
Case-sensitive finding by name is rarely useful, and on the other hand,
substring search made in the fuzzy mode is not always preferable even when
case-insensitivity is required.

With this change, TrackBusinessLayer::findAllByNameAndArtistName no longer
needs to use the fuzzy mode, and it can be simplified. This function is used
only when fetching track lyrics from the Subsonic API.