Match each whitespace-separated substring separately on fuzzy findAllByName
When BaseMapper::findAllByName is called with argument $fuzzy=true and a
searched string containing any whitespace characters, the default behavior is
now to split the string by the whitespace and search each substring
separately but in the given order from the name field.

The old behavior of treating the whole input as a single substring can be
requested by enclosing the given input in quotation marks (").

Among other things, this fixes the search on the Substreamer client which
implicitly adds the quotation marks to each search query given by the user.