Fix continuing playback from the same offset on navigation from Files to Music
When clicking the song title in the "lite" player within Files, the playback
was supposed to continue from the same time offset within the Music app
(since v0.16.0). However, this had been broken by the commit 797a01ee
included in v1.0.0, and the playback always started again from the beginning
of the track.