Fix misleading error when viewing details for an album not found from Last.fm
In this particular case, Last.fm returns the status code 404. This was
incorrectly interpreted as a sign of broken Last.fm connection and the error
message suggested that maybe the API key is incorrect. With this commit, the
status code 404 is give a special treatment so that it is not considered a
connectivity issue.