Subsonic: Fix arg `musicFolderId` on `getIndexes` not being optional
The argument `musicFolderId` of the method `getIndexes` ought to be
optional according to the API specification, but it was mistakenly
marked as "required" on the commit 4f323083. This broke the
compatibility with the Soundwaves Player and prevented using the "All
Folders" selection on many other players, e.g. DSub and Subsonic Music
Streamer.

refs #885