Set the substitute character to the default value ('?') for getID3
The getID3 library implicitly assumes that the global substitute
character setting is in its default value 0x3F, i.e. a question mark.
This is relevant when it evaluates, if an id3v1 or riff tag is actually
just converted version of the more proper data in id3v2, and
non-ISO-Latin-1 characters have been substituted with question marks.

Previously, it depended on the use case, which value this global
setting had when running getID3: it was different thing to scan files
than it was to view file details. Now we ensure systematic behavior in
all cases.