Fix migration on NC from Music app older than v1.0.0
When defining the DB schema for Nextcloud was moved to use the
migrations mechanism in v1.2.1 (commit 6982d91b), a problem was
introduced: The migration worked only either on a clean database (no
earlier versions of Music installed) or on v1.0.0 or newer. Installing
over any v0.x.y got broken. This was because the auto-generated
migration script Version010200Date20210513171803 newer altered the DB
schema of existing tables; it just created tables which were missing.

The not so well-working migration script has now been replaced with the
hand-written Version010000Date20210903000000. This script checks
separately that each needed column exists in the database and creates
the missing ones. It also removes couple of obsolete columns (removed
over the years) if those are still present in the database.

This new script should be able to migrate the database from any version
starting from v0.4.0 to the v1.0.0 level. Most of these versions have
not been tested, though.

refs https://github.com/owncloud/music/issues/889
refs https://github.com/owncloud/music/issues/883