Ampache: Support type `podcast` also on the action `stream`
This kind of support was already there on the action `download` but it
was accidentally left out from the action `stream` which is basically
identical action on our implementation.

Also, we now support both types `podcast` and `podcast_episode` on both
of these actions. On both cases, the given `id` is interpreted as a
podcast episode ID. The API spec v4 used here the type value `podcast`
while API spec v4 uses more consistent `podcast_episode`.

refs https://github.com/owncloud/music/issues/891