Use 2nd level parent folder of a track as fallback for the artist name
In case the audio file has no metadata tag for the artist name or this
cannot be extracted for some reason, the name of the parent folder of
the parent folder of the file is now used as a fallback name.

If the file has not two or more levels of parent folders (in addition
to the user root folder), then the artist name in the mentioned case is
still set to "Unknown artist" as before.