One more attempt to make the NC database migrations correctly
The commit ff96ad2a6 caused a new problem when updating on Nextcloud
from Music v1.3.1 to v1.3.2. It probably depended on the DBMS used
and/or the contents of the previous database, if this prevented the
update altogether. But even if it succeeded, the DB schema for the table
oc_music_bookmarks got messed up (the table ended up containing one
obsolete column and one obsolete unique index which were not supposed to
be there after v1.3.0).

In this commit, the functionality of two previous migration scripts is
combined to a single script file. This makes the overall logic easier
to follow and should be less prone for surprises.

refs https://github.com/owncloud/music/issues/892