Fix German translations being used only for the "Formal: Sie" variant
The language codes in the beginning of the other German language
variants were incorrectly copy-pasted from the de_DE variant, which is
used when the user has selected the language as "Deutsch (Förmlich:
Sie)". As a result, the compiled translations.js contained four times
translations for this variant and no translations for any other German
variants.

refs https://github.com/owncloud/music/pull/890