Support playing and adding to playlist recursively from parent folder
When the tree layout is used for the Folders view, dragging any parent
folder to a playlist now adds recursively all the tracks found anywhere
under the dragged folder. Similarly, clicking the title of a folder
with sub-folders sets the play-scope to include all the tracks from all
the sub-folders.