Construct the folder paths on the front-end instead of back-end
The previous solution of getting the folder paths on the back-end
didn't work correctly in case the music library path was on an external
storage. Trying to fix that on the back-end would have been
error-prone, at best. On the other hand, we can now trivially construct
the paths on the front-end by relying on the parent links, and the time
used on this is negligible even with huge libraries. Also, this
significantly reduces the size of the response of the `/api/folders`.

Also, the paths used on the UI no longer include the path of the
library root folder. This makes sense especially when using the search,
because previously, anything matching the root folder path matched
every single track in the library.