Fix navigation pane auto-collapse on recent versions of Nextcloud
The navigation pane is supposed to auto-collapse upon navigation when
in the mobile layout. However, this didn't work on recent version of
Nextcloud (21 and 22, at least) although it did work on ownCloud and
Nextcloud 17.

The way the system has been designed to work is that the pane should be
collapsed by the core when an <a> or a <button> element is clicked
within the navigation pane. But most of our navigation pane items do
not belong to either group; this has been necessary to make the layout
including the context menus work properly on all versions of NC and OC.

Previously, we had worked around this by simulating a click on the
#app-navigation-toggle button. But for an unknown reason, this does not
work on NC21/22, although the element still does exist and manual click
on it collapses the pane. Hence, we have now implemented a different
kind of workaround: Theres is a hidden fake button within the
navigation pane, and we simulate click on this fake button when ever we
want to collapse the pane. This seems to work on all supported cloud
versions.