Respect the locale user setting when formatting dates and times
The locale setting from the global user settings of the cloud are now
respected when formatting dates and times shown in the details pane for
various entries. Previously, the formatting happened based on the
browser default locale which usually depends on the UI language of the
browser. User settings of NC have an actual locale selection while OC
deduces the locale from the selected UI language.

Note, that for some of the locales, the example date format shown by
Nextcloud settings is differs from what we are actually showing in
Music. This is probably because NC uses the moment.js library to format
the dates while we use Javascript native Date.toLocaleString.
Meanwhile, NC has deprecated the use of the global moment.js, so
relying on that wouldn't be wise. And as this library is no longer
maintained, there's actually a good chance that the formatting given by
JS is more proper.