Optimize the REST end point `/api/playlists/{id}?fulltree=1`
Probably this fulltree mode is not used by anyone, so the real benefits
are dubious, but the previous implementation was laughably inefficient.
Hugely more efficient solution was possible with a tad less code
because we have previously created some handy tools for this.

To make the optimized version surely work also on all installations of
SQLite, a chunked strategy was applied to BusinessLayer::findById. This
may help to prevent some other problems, too.