Use HTML5 localStorage instead of cookies to store web UI settings
The cookies are sent to the server on each request, which wasn't
necessary at all in our case: all the cookies were used to store
settings needed only on the front-end. Hence, the localStorage is more
natural choice, and it should be well supported, starting already from
IE8. This change also gets us rid of the warning in the Firefox console
about cookies without a sameSite property.

Upon the first page load with the changed SW, the previous settings are
migrated from the cookies to the localStorage and the cookies are
removed. This logic may be removed in some later version, once majority
of the users have updated their Music app. It shouldn't be a big deal
if some few users then miss this migration since losing the front-end
settings is a minor nuisance, and being stuck with the unused cookies
doesn't cause any real problem. Then, we could ditch the js-cookie
library, too.