Direct the Subsonic and Ampache base URLs to the Music app front page
At least in the server settings of DSub, there is a button "Open in
browser" (*) which opens the given server address to the web browser.
This address is something like
https://<server_address>/index.php/apps/music/subsonic. As this base
address wasn't routed to anywhere, it used to open the cloud's default
view. Now it is routed to the Music app main view.

(*) For another example of similar concept, the Jamstash client has a
"Subsonic" button in the upper right corner which is a link to the
configured base URL of the server.