Subsonic: Start browsing by folders from the library root contents
When browsing by folders on the Subsonic API, the main level returned
by `getIndexes` is now the contents of the music library root folder.
Previously, this method returned just a single folder which was the
library root folder.

The down-side of this is that several Subsonic clients (*) are unable to
show any tracks stored directly on the library root folder. This,
however, is a client issue because the API specification clearly states
that this is possible and even shows it in the example. The DSub player
shows also the root level tracks properly. And this is how the other
Subsonic servers work, too. It may be also argued, that users
interested in the browsing by folders probably don't use flat folder
structure.

(*) At least Ultrasonic, Supersonic, and the official Subsonic Android
app have this problem