Subsonic: Optimize loading the tracks of long playlists
The Subsonic method `getPlaylist` and several other methods potentially
returning a long list of tracks are now much more efficient on long
lists. This is achieved by injecting the albums to all returned tracks
at once instead of doing it one-by-one for each track.

As an example, showing a playlist of 688 songs on DSub previously took
~17 seconds and after the optimization only ~1 second.