Another attempt to fix the layout of html-formatted podcast details
The commit 00010caa already tried to fix the layout of the
html-formatted podcast details but it didn't work quite as expected:
Unsetting the `display` property doesn't really return the browser
default property value, but it sets the display to "inline" (or
inherited value). Hence, the previous commit fixed the layout of some
elements (e.g. <a>, <span>) but broke the layout of some others (e.g.
<p>).

To properly fix the issue, we need to avoid the previous overly
general rule of setting everything inside #app-sidebar to `display:
inline-block`. Instead, it is now set to only those elements which
actually benefit from it, i.e. things we don't want get split to
multiple lines: <a> elements within long descriptions and the "similar
artist" items in the artist details. The layout of these similar
artists was also improved now so that the separator comma always stays
on the same line with the pervious artist.